Extract all DB table types into dedicated files
I found this makes it a lot eaiser to quickly scan the TS project and understand all
the DB tables.
