+++ src/knex-types/tbc_account_balance.ts
@@ -1,0 +1,6 @@
+export interface TBCAccountBalance {
+  id?: number;
+  tbc_account_id: number;
+  datetime: Date;
+  balance: number;
+}
