+++ src/knex-types/token_account.ts
@@ -1,0 +1,7 @@
+export interface TokenAccount {
+  id?: number;
+  address: Uint8Array;
+  owner_address?: Uint8Array;
+  mint_id: number;
+  first_transaction_date: Date;
+}
