+++ src/knex-types/token_account_balance.ts
@@ -1,21 +1,3 @@
-// this might end up kind of redundant with TBCTokenMint or the token B info in TBCAccounts if we add those there,
-// but seemed like a better separation of concerns to have a dedicated table for this, otherwise we'd probably need
-// bool configuration flags e.g. `should_get_tbc_tvl` `should_get_token_balances` to separate use cases
-// (and seems cleaner to have this be a separate table than as a series of rows in TBCAccounts)
-export interface TokenAccountMint {
-  id?: number;
-  mint_address: Uint8Array;
-  decimals: number;
-}
-
-export interface TokenAccount {
-  id?: number;
-  address: Uint8Array;
-  owner_address?: Uint8Array;
-  mint_id: number;
-  first_transaction_date: Date;
-}
-
 export interface TokenAccountBalance {
   id?: number;
   token_account_id: number;
@@ -29,12 +11,4 @@ export interface TokenAccountBalance {
   // (if we start trying to piece together whether a negative change caused the account to go to 0, might as well just
   // try to get the real balance)
   approximate_minimum_balance: number;
-}
-
-export interface TokenAccountTransaction {
-  id?: number;
-  token_account_id: number;
-  datetime: Date;
-  transaction_hash: Uint8Array;
-  transfer_in: boolean;
 }
