+++ src/knex-types/token_account_transaction.ts
@@ -1,0 +1,7 @@
+export interface TokenAccountTransaction {
+  id?: number;
+  token_account_id: number;
+  datetime: Date;
+  transaction_hash: Uint8Array;
+  transfer_in: boolean;
+}
